refactor: AsyncAPISpecificationReferencesCollector move from _core to extensions.index
